fixed csp check

+++ install.sh
@@ -27,14 +27,16 @@ fi
 
 # check to see if VM belongs to one of the big 3 CSPs
 # done via checking against https://ipinfo.io/, getting the org and cutting out the asn
+# then checking to see if the metadata server/internal address responds
 IPINFO_ORG=$(curl -s ipinfo.io/org | awk {'first = $1; $1=""; print $0'}|sed 's/^ //g')
-if [ 'Google LLC' = "${IPINFO_ORG}" ]
+
+if [[ 'Google LLC' = "${IPINFO_ORG}" && $(curl --write-out %{http_code} -s --output /dev/null 169.254.169.254) == '200' ]]
 then
   cp ./00_gcp.sh ${DYNMOTD_CUSTOM_SCRIPTS_PATH}/.
-elif [ 'Amazon.com, Inc.' = "${IPINFO_ORG}" ]
+elif [ 'Amazon.com, Inc.' = "${IPINFO_ORG}" && $(curl --write-out %{http_code} -s --output /dev/null 169.254.169.254) == '200' ]]
 then
   cp ./00_aws.sh ${DYNMOTD_CUSTOM_SCRIPTS_PATH}/.
-elif [ 'Microsoft Corporation' = "${IPINFO_ORG}" ]
+elif [ 'Microsoft Corporation' = "${IPINFO_ORG}" && $(curl --write-out %{http_code} -s --output /dev/null 169.254.169.254) == '400' ]]
 then
   cp ./00_azure.sh ${DYNMOTD_CUSTOM_SCRIPTS_PATH}/.
 fi
